Admin refinements: route-group sub-pages, primary button style, form chrome
- Tabbed pages move into a route group at `src/app/admin/(tabbed)/` so they
  inherit the tabs layout, while drill-in sub-pages (e.g. `/admin/clients/new`,
  `/admin/evaluations/new`, `/admin/evaluations/[id]`) live outside the group
  and don't. The new top-level `src/app/admin/layout.tsx` keeps the admin
  auth gate; the tabs/header live in `(tabbed)/layout.tsx`.

- Add Client and New Evaluation render their own page chrome — a "← Clients" /
  "← Evaluations" back link, an h1, and a card-wrapped form — matching the
  team-member detail pattern. New Evaluation expands to the page max-width and
  the form is wrapped in a card to match Add Client.

- Primary buttons on the Clients, Evaluations, and Beta Codes tabs (and their
  forms) move to the established style: square (no rounded corners), uppercase
  mono with widest tracking, ladder-green background, dark text. Matches the
  buttons on /settings, /score, and the recently-restyled team setup banner.

- Evaluations detail/report drill-ins (/admin/evaluations/[id], /report) keep
  their existing buttons for now; per-section refinement will normalize them.
